warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@11939 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/BzMaterial.cxx
@@ -587,7 +587,7 @@ void BzMaterial::print(std::ostream& out, const std::string& indent) const
     out << "Ns " << (1000.0f * (shininess / 128.0f)) << std::endl;
     if (textureCount > 0) {
       const TextureInfo* ti = &textures[0];
-      const unsigned int nlen = ti->name.size();
+      const unsigned int nlen = (unsigned int)ti->name.size();
       if (nlen > 0) {
         std::string texname = ti->name;
         const char* cname = texname.c_str();
